requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via django
attrs==21.2.0
    # via pytest
coverage==5.5
    # via pytest-cov
django==3.2.5
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
importlib-metadata==4.6.1
    # via
    #   pluggy
    #   pytest
iniconfig==1.1.1
    # via pytest
packaging==21.0
    # via pytest
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pyparsing==2.4.7
    # via packaging
pytest==6.2.4
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-stress
pytest-cov==2.12.1
    # via -r requirements.in
pytest-django==4.4.0
    # via -r requirements.in
pytest-stress==1.0.1
    # via -r requirements.in
pytz==2021.1
    # via django
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via
    #   pytest
    #   pytest-cov
typing-extensions==3.10.0.0
    # via
    #   asgiref
    #   importlib-metadata
zipp==3.5.0
    # via importlib-metadata
